Added flttr init command

+++ lib/commands/buildable.rb
@@ -5,10 +5,9 @@ require 'fileutils'
 
 module Commands
     class Buildable   
-        attr_accessor :verbose, :platform, :flavor, :release    
+        attr_accessor :platform, :flavor, :release    
 
         def initialize(args)
-            @verbose = args[:verbose]
             @platform = args[:platform]
             @flavor = args[:flavor]
             @release = args[:release]
@@ -17,8 +16,8 @@ module Commands
         def execute
             # Check if the directory contains a pubspec.yaml file, which is required for a Flutter project.
             unless File.exist? "pubspec.yaml"
-                warn colored :red, " [!] The directory doesn't contain a pubspec.yaml, make sure to run this command inside a Flutter project. " 
-                exit
+                warn colored :red, "#{CHAR_ERROR} The directory doesn't contain a pubspec.yaml, make sure to run this command inside a Flutter project. " 
+                # exit
             end
 
             prepare_dir
@@ -33,18 +32,22 @@ module Commands
         end
 
         def prepare_dir
+            puts colored :default, "\n#{CHAR_VERBOSE} Checking for existence of config/ dir" unless !$verbose
+
             # Creating config directory if it doesn't exist
             Dir.mkdir "config" unless File.exist? "config"
         end
 
         def copy_configuration_files
             begin
+                puts colored :default, "\n#{CHAR_VERBOSE} Loading config/.config.yaml contents" unless !$verbose
                 configuration = YAML.load_file('config/.config.yaml')
             rescue
+                warn colored :yellow, "\n#{CHAR_WARNING} File config/config.yaml doesn\'t exist, using empty configuration: {}"
                 configuration = {}
             end
 
-            puts colored :blue, "\n[:] Copying configuration files"
+            puts colored :blue, "\n#{CHAR_CHECK} Copying configuration files"
 
             if configuration.key?('files')
                 filesToCopy = configuration['files']
@@ -63,8 +66,10 @@ module Commands
         def write_build_config
             begin
                 # Build a .build.json file
+                puts colored :default, "\n#{CHAR_VERBOSE} Loading config/#{@flavor}.json" unless !$verbose
                 buildConfig = JSON.load(File.open("config/#{@flavor}.json"))
             rescue
+                warn colored :yellow, "\n#{CHAR_WARNING} File config/#{@flavor}.json doesn't exit, using empty configuration: {}" 
                 buildConfig = {}
             end
 
@@ -72,9 +77,10 @@ module Commands
 
             begin
                 # Open signing configuration
+                puts colored :default, "\n#{CHAR_VERBOSE} Loading #{platformConfigFileName}" unless !$verbose
                 platformConfigAll = JSON.load(File.open(platformConfigFileName))
             rescue
-                warn colored :red, "\n[!] File #{platformConfigFileName} does not exist, falling back to default {}\n\n"
+                warn colored :yellow, "\n#{CHAR_WARNING} File #{platformConfigFileName} does not exist, using empty configuration: {}"
                 platformConfigAll = {}
             end
 
@@ -82,17 +88,15 @@ module Commands
                 platformConfig = platformConfigAll["default"] || {}
                 platformConfig = platformConfig.merge(@release ? platformConfigAll["release"] || {} : platformConfigAll["debug"] || {})
 
-                if :verbose
-                    puts colored :blue, "\n[:] Loaded platform config, combining it to the default config"
-                    pp platformConfig
-                end
+                puts colored :green, "\n#{CHAR_CHECK} Loaded platform config, combining it to the default config"
+
+                pp platformConfig unless !$verbose
                 
                 buildConfig = buildConfig.merge(platformConfig)
             else
-                if @verbose
-                    puts colored :blue, "\n[:] Loaded platform config, combining it to the default config"
-                    pp platformConfigAll
-                end
+                puts colored :green, "\n#{CHAR_CHECK} Loaded platform config, combining it to the default config"
+
+                pp platformConfigAll unless !$verbose
 
                 buildConfig = buildConfig.merge(platformConfigAll)
             end
@@ -100,6 +104,9 @@ module Commands
             File.open('./config/.build.json', 'w') do |file|
                 file.write(JSON.dump(buildConfig))
             end
+
+            puts colored :green, "\n#{CHAR_CHECK} Wrote configuration to config/.build.json"
+            pp buildConfig unless !$verbose
 
             return buildConfig
         end
@@ -110,9 +117,16 @@ module Commands
             # We need specific values to be updated before starting flutters own build script.
             xcconfigFileName = 'ios/Flutter/Generated.xcconfig'
 
+            unless File.exist? "ios"
+                warn colored :yellow, "\n#{CHAR_WARNING} No iOS folder is present in the directory, make sure you're running the command from inside a projects root" 
+                return
+            end
+
             Dir.mkdir "ios/Flutter" unless File.exist? "ios/Flutter"
 
             generatedXcodeConfig = {}
+
+            puts colored :default, "\n#{CHAR_VERBOSE} Loading the values from #{xcconfigFileName}" unless !$verbose
 
             if File.exist?(xcconfigFileName)
                 File.open(xcconfigFileName) do |file|
@@ -123,23 +137,21 @@ module Commands
                 end
             end
 
+            puts colored :default, "\n#{CHAR_VERBOSE} Overriding the values in Generated.xcconfig to the build configuration" unless !$verbose
+
             # Now overwrite the existing keys with our noewly created buildConfig
             for key in buildConfig.keys
                 generatedXcodeConfig[key] = buildConfig[key]
             end
 
-            if :verbose
-                puts colored :blue, "\n[:] Writing to Generated.xcconfig"
-            end
+            puts colored :blue, "\n#{CHAR_FLAG} Writing to Generated.xcconfig" unless !$verbose±
 
             # Now print all lines back into Generated.xcconfig
             File.open(xcconfigFileName, "w") do |file|
                 for key in generatedXcodeConfig.keys
                     line = "#{key}#{key.start_with?("//") ? "" : "=#{generatedXcodeConfig[key]}"}"
 
-                    if :verbose
-                        puts " - Writing into Generated.xcconfig: #{line}"
-                    end
+                    puts colored :default, "\n#{CHAR_VERBOSE} Writing into Generated.xcconfig: #{line}" unless !$verbose
 
                     file.write "#{line}\n"
                 end
